readyup|refactor: Move the kits' tests and helpers beside the kit code
The kits' tests lived at the package root under `__tests__/kits/` and their helpers under `test-utils/kits/`, so every test reached its helpers with a `../../` hop into a sibling tree, and neither `.readyup/kits/` nor `.readyup/kits/checks/` had a `__tests__/` sibling to show what was covered. Both now sit inside the tree they serve: a `__tests__/` beside each directory holding tested code, and `test-utils/` beside the tests that import it.

A nested `.npmignore` keeps them out of the tarball. The `files` allowlist ships `.readyup` wholesale, because the publishing kit's `describeFilesCoverage` accepts only an entry naming the directory or the package root; an ignore file nested inside an included directory still prunes, where a package-root one is overridden by `files`.

Coverage collapses to a single glob: the helpers now fall under `.readyup/kits/**`, and the inherited exclude drops the `__tests__/` directories that glob would otherwise sweep in.
