+++ packages/readyup/.readyup/kits/__tests__/publishing.test.ts
@@ -5,15 +5,10 @@ import process from 'node:process';
 
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
-import type { RdyResult } from '../../src/types.ts';
-import { pickResult, runChecklist } from '../../test-utils/kits/checklist-results.ts';
-import { loadOwnKit } from '../../test-utils/kits/loadOwnKit.ts';
-import {
-  FIXTURE_KITS_DIR,
-  writeKit,
-  writeKitManifest,
-  writePackageJson,
-} from '../../test-utils/kits/project-fixture.ts';
+import type { RdyResult } from '../../../src/types.ts';
+import { pickResult, runChecklist } from '../test-utils/checklist-results.ts';
+import { loadOwnKit } from '../test-utils/loadOwnKit.ts';
+import { FIXTURE_KITS_DIR, writeKit, writeKitManifest, writePackageJson } from '../test-utils/project-fixture.ts';
 
 /** Bundle reaching for a package that only the publishing project has installed. */
 const LEAKY_BUNDLE = 'import picomatch from "picomatch";\nexport default { checklists: [] };\n';
