+++ packages/readyup/tsconfig.json
@@ -5,7 +5,6 @@
     "src/**/*.ts",
     "src/**/*.d.ts",
     "__tests__/",
-    "test-utils/",
     ".config/**/*.ts",
     ".readyup/kits/**/*.ts",
     "config/",
