+++ packages/readyup/vitest.config.ts
@@ -14,11 +14,13 @@ import { defineVitestConfig } from '@williamthorsen/nmr/vitest';
 // does, without a prior build standing between the test and the source. Longest specifier first: an alias on `readyup`
 // also matches `readyup/check-utils`, and the first match wins.
 //
-// The coverage entries extend the inherited `src`-only glob, which would otherwise leave both trees unmeasured. Kits
-// are authored under `.readyup/` rather than `src/` because that is the directory readyup itself reads, and helpers
-// under `test-utils/` are measured on purpose: an uncovered helper is why they sit outside `__tests__/` at all.
-// The text reporter prints no `test-utils` group even so; the files are in the coverage data, as the json and html
-// reporters show, so this is a reporter display gap rather than a glob that failed to match.
+// The coverage entry extends the inherited `src`-only glob, which would otherwise leave the kit tree unmeasured. Kits
+// are authored under `.readyup/` rather than `src/` because that is the directory readyup itself reads. One glob
+// covers the checks and the helpers alike, since `test-utils/` sits inside the tree it serves; the inherited exclude
+// drops the `__tests__/` directories the same glob would otherwise sweep in. Helpers are measured on purpose: an
+// uncovered helper is why they sit outside `__tests__/` at all. The text reporter prints no `test-utils` group even
+// so; the files are in the coverage data, as the json and html reporters show, so this is a reporter display gap
+// rather than a glob that failed to match.
 export default defineVitestConfig({
   project: { env: { RDY_STYLE: 'rich' } },
   root: {
@@ -30,7 +32,7 @@ export default defineVitestConfig({
     },
     test: {
       coverage: {
-        include: ['.readyup/kits/**/*.ts', 'test-utils/**/*.ts'],
+        include: ['.readyup/kits/**/*.ts'],
       },
     },
   },
